+++ ManageAzure/AzureModels/VirtualMachine.cs
@@ -8,11 +8,11 @@ namespace ManageAzure.AzureModels
         public string RoleName { get; set; }
         public string RoleSize { get; set; }
         public string RoleType { get; set; }
-        public int HourlyRate { get; set; }
-        public int MonthlyRate { get; set; }
+        public double HourlyRate { get; set; }
+        public double MonthlyRate { get; set; }
 
         public VirtualMachine() { }
-        public VirtualMachine(string roleName, string roleSize, string roleType, int hourlyRate) 
+        public VirtualMachine(string roleName, string roleSize, string roleType, double hourlyRate) 
         {
             RoleName = roleName;
             RoleSize = roleSize;
